+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-## 28-01-2024
+## History of changes on project
 
-- [Commit-1]
+- [Commit-2]/ 30-01-2024
+
+  - Added Redux, react-redux to app
+
+- [Commit-1]/ 28-01-2024
+
   - Created all fundament of App
   - Added Reducers ( with business logic of Tasks and TodoLists) and used it with hook UseReducer
   - Added unit tests for reducers